requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
simple-term-menu==1.6.6
readchar==4.1.0
Pillow==10.4.0